requirements.txt
streamlit
tensorflow
numpy
pillow
psycopg2-binary
